Vaporeon
Hey guys, did you know that in terms of Pokémon, Vaporeon is the most compatible Pokémon for human friendships? Not only are they an Eeveelution, which is mostly comprised of domesticable mammals, Vaporeon are an average of 3'03' tall and 63.9 pounds. this means they're small enough to be able sit comfortably on human laps, and with their impressive Base Stats for HP and access to Acid Armor, you can be sure that they will stay safe and be resistant to illness and danger. Due to their mostly water based biology, there's no doubt in my mind that a Vaporeon would be incredibly wet, so wet that, in the case of a house fire -- likely quite common in the Pokémon universe, considering the number of fire-types -- you can feel safe in the knowledge that they will be able to remain safe until fire services arrive. They can also learn the moves Attract, Baby-Doll Eyes, Captivate, and Charm, along with not having fur, so it'd be incredibly easy to convince strict landlords that they won't cause any problems. With their abilities Water Absorb and Hydration, they make for low-maintenance pets with enough water. No other Pokémon comes close to this level of compatibility. Also, fun fact, if you're lucky enough, you may even find a purple Vaporeon.